[fix] 코드리뷰 반영 TipId, SpotId URL에서 빼고 body 로 같이 보내기
