+++ app/api/admin/tips/[tipId]/route.ts
@@ -1,6 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
 import GetReactionUsecase from "@/application/usecases/spot/tips/GetReactionUsecase";
-import DeleteTipUsecase from "@/application/usecases/spot/tips/DeleteTipUsecase";
 import { PgTipRepository } from "@/infrastructure/repositories/PgTipRepository";
 import { PgImageRepository } from "@/infrastructure/repositories/PgImageRepository";
 import PgSpotRepository from "@/infrastructure/repositories/PgSpotRepository";
@@ -49,33 +48,6 @@ export async function GET(req: NextRequest) {
     console.error("❌ GET 요청 처리 중 오류 발생:", error);
     return NextResponse.json(
       { error: "서버 오류가 발생했습니다." },
-      { status: 500 }
-    );
-  }
-}
-
-export async function DELETE(request: Request) {
-  try {
-    const { tipId, spotId } = await request.json();
-    const tipRepository = new PgTipRepository();
-    const spotRepository = new PgSpotRepository();
-
-    const deleteTipUsecase = new DeleteTipUsecase(
-      tipRepository,
-      spotRepository
-    );
-
-    await deleteTipUsecase.execute(tipId, spotId);
-
-    return NextResponse.json(
-      { message: "Tip deleted successfully" },
-      { status: 200 }
-    );
-  } catch (error) {
-    console.error("Error deleting tip:", error);
-
-    return NextResponse.json(
-      { error: "Failed to delete tip" },
       { status: 500 }
     );
   }
