+++ app/api/admin/tips/route.ts
@@ -2,6 +2,7 @@ import { NextResponse } from "next/server";
 import { GetTipListUseCase } from "@/application/usecases/admin/tip/GetTipListUseCase";
 import { PgTipRepository } from "@/infrastructure/repositories/PgTipRepository";
 import PgSpotRepository from "@/infrastructure/repositories/PgSpotRepository";
+import DeleteTipUsecase from "@/application/usecases/spot/tips/DeleteTipUsecase";
 
 export async function GET() {
   try {
@@ -16,6 +17,33 @@ export async function GET() {
     console.error("Error fetching tips:", error);
     return NextResponse.json(
       { error: "Failed to fetch tips" },
+      { status: 500 }
+    );
+  }
+}
+
+export async function DELETE(request: Request) {
+  try {
+    const { tipId, spotId } = await request.json();
+    const tipRepository = new PgTipRepository();
+    const spotRepository = new PgSpotRepository();
+
+    const deleteTipUsecase = new DeleteTipUsecase(
+      tipRepository,
+      spotRepository
+    );
+
+    await deleteTipUsecase.execute(tipId, spotId);
+
+    return NextResponse.json(
+      { message: "Tip deleted successfully" },
+      { status: 200 }
+    );
+  } catch (error) {
+    console.error("Error deleting tip:", error);
+
+    return NextResponse.json(
+      { error: "Failed to delete tip" },
       { status: 500 }
     );
   }
